+++ src/main/java/com/e2i/wemeet/service/team/TeamServiceImpl.java
@@ -11,27 +11,18 @@ import com.e2i.wemeet.domain.profileimage.ProfileImageRepository;
 import com.e2i.wemeet.domain.team.Team;
 import com.e2i.wemeet.domain.team.TeamRepository;
 import com.e2i.wemeet.domain.team.invitation.InvitationAcceptStatus;
-import com.e2i.wemeet.domain.team.invitation.TeamInvitation;
 import com.e2i.wemeet.domain.team.invitation.TeamInvitationRepository;
 import com.e2i.wemeet.domain.teampreferencemeetingtype.TeamPreferenceMeetingType;
 import com.e2i.wemeet.domain.teampreferencemeetingtype.TeamPreferenceMeetingTypeRepository;
 import com.e2i.wemeet.dto.request.team.CreateTeamRequestDto;
-import com.e2i.wemeet.dto.request.team.InviteTeamRequestDto;
 import com.e2i.wemeet.dto.request.team.ModifyTeamRequestDto;
 import com.e2i.wemeet.dto.response.team.MyTeamDetailResponseDto;
 import com.e2i.wemeet.dto.response.team.TeamManagementResponseDto;
 import com.e2i.wemeet.dto.response.team.TeamMemberResponseDto;
-import com.e2i.wemeet.exception.badrequest.GenderNotMatchException;
-import com.e2i.wemeet.exception.badrequest.InvitationAlreadyExistsException;
-import com.e2i.wemeet.exception.badrequest.InvitationAlreadySetException;
 import com.e2i.wemeet.exception.badrequest.ManagerSelfDeletionException;
 import com.e2i.wemeet.exception.badrequest.NonTeamMemberException;
 import com.e2i.wemeet.exception.badrequest.NotBelongToTeamException;
-import com.e2i.wemeet.exception.badrequest.TeamAlreadyActiveException;
-import com.e2i.wemeet.exception.badrequest.TeamAlreadyExistsException;
-import com.e2i.wemeet.exception.notfound.InvitationNotFoundException;
 import com.e2i.wemeet.exception.notfound.MemberNotFoundException;
-import com.e2i.wemeet.exception.unauthorized.UnAuthorizedUnivException;
 import jakarta.servlet.http.HttpServletResponse;
 import java.security.SecureRandom;
 import java.util.ArrayList;
@@ -59,7 +50,6 @@ public class TeamServiceImpl implements TeamService {
     public Long createTeam(Long memberId, CreateTeamRequestDto createTeamRequestDto,
         List<Code> teamPreferenceMeetingTypeList, HttpServletResponse response) {
         Member member = findMember(memberId);
-        validateMember(member);
 
         // team 생성
         Team team = teamRepository.save(
@@ -113,49 +103,6 @@ public class TeamServiceImpl implements TeamService {
             .managerImageAuth(team.getMember().isImageAuth())
             .preferenceMeetingTypeList(preferenceMeetingTypeToCodeString(preferenceMeetingTypeList))
             .build();
-    }
-
-    @Transactional
-    @Override
-    public void inviteTeam(Long memberId, InviteTeamRequestDto inviteTeamRequestDto) {
-        Member teamMember = memberRepository.findByNicknameAndMemberCode(
-                inviteTeamRequestDto.nickname(),
-                inviteTeamRequestDto.memberCode())
-            .orElseThrow(MemberNotFoundException::new);
-
-        Member manager = findMember(memberId);
-        Team team = manager.getTeam();
-
-        validateInvitation(teamMember, manager, team);
-
-        teamInvitationRepository.save(
-            TeamInvitation.builder()
-                .acceptStatus(InvitationAcceptStatus.WAITING)
-                .member(teamMember)
-                .team(team)
-                .build());
-    }
-
-    @Transactional
-    @Override
-    public void takeAcceptStatus(Long memberId, Long invitationId, boolean accepted) {
-        Member member = findMember(memberId);
-        validateMember(member);
-
-        TeamInvitation teamInvitation =
-            teamInvitationRepository.findByTeamInvitationIdAndMemberMemberId(
-                    invitationId, memberId)
-                .orElseThrow(InvitationNotFoundException::new);
-
-        if (teamInvitation.getAcceptStatus() != InvitationAcceptStatus.WAITING) {
-            throw new InvitationAlreadySetException();
-        }
-
-        if (accepted) {
-            acceptInvitation(teamInvitation);
-        } else {
-            rejectInvitation(teamInvitation);
-        }
     }
 
     @Transactional(readOnly = true)
@@ -236,7 +183,6 @@ public class TeamServiceImpl implements TeamService {
             .toList();
     }
 
-
     /*
      * 팀원 목록 조회
      */
@@ -260,51 +206,6 @@ public class TeamServiceImpl implements TeamService {
             .toList();
     }
 
-    private void acceptInvitation(TeamInvitation teamInvitation) {
-        Team team = teamInvitation.getTeam();
-        if (isActiveTeam(team)) {
-            throw new TeamAlreadyActiveException();
-        }
-
-        if (team.getMembers().size() + 1 == team.getMemberCount()) {
-            team.setActive(true);
-        }
-        teamInvitation.updateAcceptStatus(InvitationAcceptStatus.ACCEPT);
-        team.setMember(teamInvitation.getMember());
-    }
-
-    private void rejectInvitation(TeamInvitation teamInvitation) {
-        teamInvitation.updateAcceptStatus(InvitationAcceptStatus.REJECT);
-    }
-
-    /*
-     * 팀 생성 및 수락이 가능한 사용자인지 확인
-     */
-    private void validateMember(Member member) {
-        if (isTeamExist(member)) {
-            throw new TeamAlreadyExistsException();
-        }
-
-        if (isUnivAuth(member)) {
-            throw new UnAuthorizedUnivException();
-        }
-    }
-
-    /*
-     * 초대 신청이 가능한지 확인
-     */
-    private void validateInvitation(Member teamMember, Member manager, Team team) {
-        // 초대한 사용자와 성별이 다른 경우
-        if (manager.getGender() != teamMember.getGender()) {
-            throw new GenderNotMatchException();
-        }
-
-        // 이미 초대한 사용자인 경우
-        if (isInvitationExist(teamMember, team)) {
-            throw new InvitationAlreadyExistsException();
-        }
-    }
-
     private List<String> preferenceMeetingTypeToCodeString(
         List<TeamPreferenceMeetingType> preferenceMeetingTypeList) {
         return preferenceMeetingTypeList.stream()
@@ -322,21 +223,6 @@ public class TeamServiceImpl implements TeamService {
 
     private boolean isTeamExist(Member member) {
         return member.getTeam() != null;
-    }
-
-    private boolean isUnivAuth(Member member) {
-        return member.getCollegeInfo().getMail() == null;
-    }
-
-    private boolean isInvitationExist(Member member, Team team) {
-        return teamInvitationRepository.findByMemberMemberIdAndTeamTeamIdAndAcceptStatus(
-                member.getMemberId(),
-                team.getTeamId(), InvitationAcceptStatus.WAITING)
-            .isPresent();
-    }
-
-    private boolean isActiveTeam(Team team) {
-        return team.isActive();
     }
 
     private void savePreferenceMeetingType(Team team, List<Code> codeList) {
